LPS-116044 Rename
