+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/AccountGroupAccountEntriesScreenNavigationCategory.java
@@ -15,17 +15,12 @@
 package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
-import com.liferay.account.admin.web.internal.security.permission.resource.AccountPermission;
-import com.liferay.account.constants.AccountActionKeys;
-import com.liferay.account.constants.AccountPortletKeys;
+import com.liferay.account.admin.web.internal.display.AccountGroupDisplay;
+import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
-import com.liferay.portal.kernel.service.permission.UserPermissionUtil;
 
 import java.io.IOException;
 
@@ -41,15 +36,19 @@ import org.osgi.service.component.annotations.Reference;
  * @author Albert Lee
  */
 @Component(
-	property = "screen.navigation.entry.order:Integer=20",
-	service = ScreenNavigationEntry.class
+	property = {
+		"screen.navigation.category.order:Integer=20",
+		"screen.navigation.entry.order:Integer=10"
+	},
+	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
 )
-public class AccountUserAccountsScreenNavigationEntry
-	implements ScreenNavigationEntry<User> {
+public class AccountGroupAccountEntriesScreenNavigationCategory
+	implements ScreenNavigationCategory,
+			   ScreenNavigationEntry<AccountGroupDisplay> {
 
 	@Override
 	public String getCategoryKey() {
-		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_GENERAL;
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ACCOUNTS;
 	}
 
 	@Override
@@ -57,37 +56,26 @@ public class AccountUserAccountsScreenNavigationEntry
 		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ACCOUNTS;
 	}
 
-	public String getJspPath() {
-		return "/account_users_admin/account_user/account_entries.jsp";
-	}
-
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(
-			locale, AccountScreenNavigationEntryConstants.ENTRY_KEY_ACCOUNTS);
+		return LanguageUtil.get(locale, "accounts");
 	}
 
 	@Override
 	public String getScreenNavigationKey() {
 		return AccountScreenNavigationEntryConstants.
-			SCREEN_NAVIGATION_KEY_ACCOUNT_USER;
+			SCREEN_NAVIGATION_KEY_ACCOUNT_GROUP;
 	}
 
 	@Override
-	public boolean isVisible(User user, User selUser) {
-		PermissionChecker permissionChecker =
-			PermissionCheckerFactoryUtil.create(user);
+	public boolean isVisible(
+		User user, AccountGroupDisplay accountGroupDisplay) {
 
-		if (AccountPermission.contains(
-				permissionChecker, AccountPortletKeys.ACCOUNT_USERS_ADMIN,
-				AccountActionKeys.ASSIGN_ACCOUNTS) ||
-			UserPermissionUtil.contains(
-				permissionChecker, selUser.getUserId(), ActionKeys.UPDATE)) {
-
-			return true;
+		if (accountGroupDisplay == null) {
+			return false;
 		}
 
-		return false;
+		return true;
 	}
 
 	@Override
@@ -97,7 +85,8 @@ public class AccountUserAccountsScreenNavigationEntry
 		throws IOException {
 
 		jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse, getJspPath());
+			httpServletRequest, httpServletResponse,
+			"/account_groups_admin/account_group/view_account_entries.jsp");
 	}
 
 	@Reference
