+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/AccountUserAccountEntriesScreenNavigationEntry.java
@@ -15,12 +15,17 @@
 package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
-import com.liferay.account.admin.web.internal.display.AccountGroupDisplay;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
+import com.liferay.account.admin.web.internal.security.permission.resource.AccountPermission;
+import com.liferay.account.constants.AccountActionKeys;
+import com.liferay.account.constants.AccountPortletKeys;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
+import com.liferay.portal.kernel.service.permission.UserPermissionUtil;
 
 import java.io.IOException;
 
@@ -36,19 +41,15 @@ import org.osgi.service.component.annotations.Reference;
  * @author Albert Lee
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.entry.order:Integer=20",
+	service = ScreenNavigationEntry.class
 )
-public class AccountGroupAccountsScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<AccountGroupDisplay> {
+public class AccountUserAccountEntriesScreenNavigationEntry
+	implements ScreenNavigationEntry<User> {
 
 	@Override
 	public String getCategoryKey() {
-		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ACCOUNTS;
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_GENERAL;
 	}
 
 	@Override
@@ -56,26 +57,37 @@ public class AccountGroupAccountsScreenNavigationCategory
 		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ACCOUNTS;
 	}
 
+	public String getJspPath() {
+		return "/account_users_admin/account_user/account_entries.jsp";
+	}
+
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "accounts");
+		return LanguageUtil.get(
+			locale, AccountScreenNavigationEntryConstants.ENTRY_KEY_ACCOUNTS);
 	}
 
 	@Override
 	public String getScreenNavigationKey() {
 		return AccountScreenNavigationEntryConstants.
-			SCREEN_NAVIGATION_KEY_ACCOUNT_GROUP;
+			SCREEN_NAVIGATION_KEY_ACCOUNT_USER;
 	}
 
 	@Override
-	public boolean isVisible(
-		User user, AccountGroupDisplay accountGroupDisplay) {
+	public boolean isVisible(User user, User selUser) {
+		PermissionChecker permissionChecker =
+			PermissionCheckerFactoryUtil.create(user);
 
-		if (accountGroupDisplay == null) {
-			return false;
+		if (AccountPermission.contains(
+				permissionChecker, AccountPortletKeys.ACCOUNT_USERS_ADMIN,
+				AccountActionKeys.ASSIGN_ACCOUNTS) ||
+			UserPermissionUtil.contains(
+				permissionChecker, selUser.getUserId(), ActionKeys.UPDATE)) {
+
+			return true;
 		}
 
-		return true;
+		return false;
 	}
 
 	@Override
@@ -85,8 +97,7 @@ public class AccountGroupAccountsScreenNavigationCategory
 		throws IOException {
 
 		jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/account_groups_admin/account_group/view_account_entries.jsp");
+			httpServletRequest, httpServletResponse, getJspPath());
 	}
 
 	@Reference
